@@ -1,6 +1,6 @@
 module github.com/rl404/akatsuki
 
-go 1.20
+go 1.21
 
 require (
 	github.com/go-chi/chi/v5 v5.0.10
