@@ -23,11 +23,11 @@ import (
 	studioSQL "github.com/rl404/akatsuki/internal/domain/studio/repository/sql"
 	"github.com/rl404/akatsuki/internal/service"
 	"github.com/rl404/akatsuki/internal/utils"
-	"github.com/rl404/fairy/cache"
+	"github.com/rl404/akatsuki/pkg/cache"
+	"github.com/rl404/akatsuki/pkg/pubsub"
 	_nr "github.com/rl404/fairy/log/newrelic"
 	nrCache "github.com/rl404/fairy/monitoring/newrelic/cache"
 	nrPS "github.com/rl404/fairy/monitoring/newrelic/pubsub"
-	"github.com/rl404/fairy/pubsub"
 )
 
 func cronFill() error {
@@ -77,7 +77,7 @@ func cronFill() error {
 	if err != nil {
 		return err
 	}
-	ps = nrPS.New(cfg.PubSub.Dialect, ps)
+	ps = nrPS.New(cfg.PubSub.Dialect, ps, nrApp)
 	utils.Info("pubsub initialized")
 	defer ps.Close()
 
@@ -119,7 +119,7 @@ func cronFill() error {
 
 	// Run cron.
 	utils.Info("filling missing data...")
-	if err := cron.New(service).Fill(nrApp, cfg.Cron.FillLimit); err != nil {
+	if err := cron.New(service, nrApp).Fill(cfg.Cron.FillLimit); err != nil {
 		return err
 	}
 
