@@ -26,11 +26,11 @@ import (
 	userAnimeSQL "github.com/rl404/akatsuki/internal/domain/user_anime/repository/sql"
 	"github.com/rl404/akatsuki/internal/service"
 	"github.com/rl404/akatsuki/internal/utils"
-	"github.com/rl404/fairy/cache"
+	"github.com/rl404/akatsuki/pkg/cache"
+	"github.com/rl404/akatsuki/pkg/pubsub"
 	_nr "github.com/rl404/fairy/log/newrelic"
 	nrCache "github.com/rl404/fairy/monitoring/newrelic/cache"
 	nrPS "github.com/rl404/fairy/monitoring/newrelic/pubsub"
-	"github.com/rl404/fairy/pubsub"
 )
 
 func cronUpdate() error {
@@ -80,7 +80,7 @@ func cronUpdate() error {
 	if err != nil {
 		return err
 	}
-	ps = nrPS.New(cfg.PubSub.Dialect, ps)
+	ps = nrPS.New(cfg.PubSub.Dialect, ps, nrApp)
 	utils.Info("pubsub initialized")
 	defer ps.Close()
 
@@ -128,7 +128,7 @@ func cronUpdate() error {
 
 	// Run cron.
 	utils.Info("updating old data...")
-	if err := cron.New(service).Update(nrApp, cfg.Cron.UpdateLimit); err != nil {
+	if err := cron.New(service, nrApp).Update(cfg.Cron.UpdateLimit); err != nil {
 		return err
 	}
 
