@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/rl404/akatsuki/internal/domain/mal/entity"
-	"github.com/rl404/akatsuki/internal/errors"
+	"github.com/rl404/fairy/errors/stack"
 	"github.com/rl404/nagato"
 )
 
@@ -26,7 +26,7 @@ func (c *Client) GetUserAnime(ctx context.Context, data entity.GetUserAnimeReque
 		),
 	)
 	if err != nil {
-		return nil, code, errors.Wrap(ctx, err)
+		return nil, code, stack.Wrap(ctx, err)
 	}
 
 	return anime, http.StatusOK, nil
