@@ -10,19 +10,9 @@ const (
 
 // Message is entity for message.
 type Message struct {
-	Type messageType `json:"type"`
-	Data []byte      `json:"data"`
-}
-
-// ParseAnimeRequest is parse anime request model.
-type ParseAnimeRequest struct {
-	ID     int64 `json:"id"`
-	Forced bool  `json:"forced"`
-}
-
-// ParseUserAnimeRequest is parse user anime request model.
-type ParseUserAnimeRequest struct {
-	Username string `json:"username"`
-	Status   string `json:"status"`
-	Forced   bool   `json:"forced"`
+	Type     messageType `json:"type"`
+	ID       int64       `json:"id"`
+	Username string      `json:"username"`
+	Status   string      `json:"status"`
+	Forced   bool        `json:"forced"`
 }
