@@ -6,6 +6,7 @@ import (
 
 	"github.com/rl404/akatsuki/internal/domain/publisher/entity"
 	"github.com/rl404/akatsuki/internal/errors"
+	"github.com/rl404/fairy/errors/stack"
 	"github.com/rl404/fairy/pubsub"
 )
 
@@ -24,34 +25,37 @@ func New(ps pubsub.PubSub, topic string) *Pubsub {
 }
 
 // PublishParseAnime to publish parse anime.
-func (p *Pubsub) PublishParseAnime(ctx context.Context, data entity.ParseAnimeRequest) error {
-	d, err := json.Marshal(data)
+func (p *Pubsub) PublishParseAnime(ctx context.Context, id int64, forced bool) error {
+	d, err := json.Marshal(entity.Message{
+		Type:   entity.TypeParseAnime,
+		ID:     id,
+		Forced: forced,
+	})
 	if err != nil {
-		return errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return stack.Wrap(ctx, err, errors.ErrInternalServer)
 	}
 
-	if err := p.pubsub.Publish(ctx, p.topic, entity.Message{
-		Type: entity.TypeParseAnime,
-		Data: d,
-	}); err != nil {
-		return errors.Wrap(ctx, errors.ErrInternalServer, err)
+	if err := p.pubsub.Publish(ctx, p.topic, d); err != nil {
+		return stack.Wrap(ctx, err, errors.ErrInternalServer)
 	}
 
 	return nil
 }
 
 // PublishParseUserAnime to publish parse user anime.
-func (p *Pubsub) PublishParseUserAnime(ctx context.Context, data entity.ParseUserAnimeRequest) error {
-	d, err := json.Marshal(data)
+func (p *Pubsub) PublishParseUserAnime(ctx context.Context, username, status string, forced bool) error {
+	d, err := json.Marshal(entity.Message{
+		Type:     entity.TypeParseUserAnime,
+		Username: username,
+		Status:   status,
+		Forced:   forced,
+	})
 	if err != nil {
-		return errors.Wrap(ctx, errors.ErrInternalServer, err)
+		return stack.Wrap(ctx, err, errors.ErrInternalServer)
 	}
 
-	if err := p.pubsub.Publish(ctx, p.topic, entity.Message{
-		Type: entity.TypeParseUserAnime,
-		Data: d,
-	}); err != nil {
-		return errors.Wrap(ctx, errors.ErrInternalServer, err)
+	if err := p.pubsub.Publish(ctx, p.topic, d); err != nil {
+		return stack.Wrap(ctx, err, errors.ErrInternalServer)
 	}
 
 	return nil
