@@ -2,12 +2,10 @@ package repository
 
 import (
 	"context"
-
-	"github.com/rl404/akatsuki/internal/domain/publisher/entity"
 )
 
 // Repository contains functions for publisher domain.
 type Repository interface {
-	PublishParseAnime(ctx context.Context, data entity.ParseAnimeRequest) error
-	PublishParseUserAnime(ctx context.Context, data entity.ParseUserAnimeRequest) error
+	PublishParseAnime(ctx context.Context, id int64, forced bool) error
+	PublishParseUserAnime(ctx context.Context, username, status string, forced bool) error
 }
