@@ -6,18 +6,14 @@ import (
 	"strings"
 
 	"github.com/rl404/akatsuki/internal/domain/mal/entity"
-	publisherEntity "github.com/rl404/akatsuki/internal/domain/publisher/entity"
 	userEntity "github.com/rl404/akatsuki/internal/domain/user_anime/entity"
-	"github.com/rl404/akatsuki/internal/errors"
+	"github.com/rl404/fairy/errors/stack"
 )
 
 // UpdateUserAnime to update user anime.
 func (s *service) UpdateUserAnime(ctx context.Context, username string) (int, error) {
-	if err := s.publisher.PublishParseUserAnime(ctx, publisherEntity.ParseUserAnimeRequest{
-		Username: strings.ToLower(username),
-		Forced:   true,
-	}); err != nil {
-		return http.StatusInternalServerError, errors.Wrap(ctx, err)
+	if err := s.publisher.PublishParseUserAnime(ctx, strings.ToLower(username), "", true); err != nil {
+		return http.StatusInternalServerError, stack.Wrap(ctx, err)
 	}
 	return http.StatusAccepted, nil
 }
@@ -39,11 +35,11 @@ func (s *service) updateUserAnime(ctx context.Context, username, status string) 
 			if code == http.StatusNotFound {
 				// Delete existing data.
 				if code, err := s.userAnime.DeleteByUsername(ctx, username); err != nil {
-					return code, errors.Wrap(ctx, err)
+					return code, stack.Wrap(ctx, err)
 				}
 				return http.StatusOK, nil
 			}
-			return code, errors.Wrap(ctx, err)
+			return code, stack.Wrap(ctx, err)
 		}
 
 		for _, a := range anime {
@@ -51,12 +47,12 @@ func (s *service) updateUserAnime(ctx context.Context, username, status string) 
 
 			// Update user anime data.
 			if code, err := s.userAnime.Update(ctx, userEntity.UserAnimeFromMal(ctx, username, a)); err != nil {
-				return code, errors.Wrap(ctx, err)
+				return code, stack.Wrap(ctx, err)
 			}
 
 			// Queue related anime.
-			if err := s.publisher.PublishParseAnime(ctx, publisherEntity.ParseAnimeRequest{ID: int64(a.Anime.ID)}); err != nil {
-				return http.StatusInternalServerError, errors.Wrap(ctx, err)
+			if err := s.publisher.PublishParseAnime(ctx, int64(a.Anime.ID), false); err != nil {
+				return http.StatusInternalServerError, stack.Wrap(ctx, err)
 			}
 		}
 
@@ -69,7 +65,7 @@ func (s *service) updateUserAnime(ctx context.Context, username, status string) 
 
 	// Delete anime not in list.
 	if code, err := s.userAnime.DeleteNotInList(ctx, username, ids, status); err != nil {
-		return code, errors.Wrap(ctx, err)
+		return code, stack.Wrap(ctx, err)
 	}
 
 	return http.StatusOK, nil
